Adjust Annex B parsing for CTS needs
The CTS relies on an external demuxer called ESExtractor. The parser
interface expects frame units to be provided, one at a time.

The best solution would be to provide a new interface that would map
the entire bitstream and depayload the trivial formats needed by the
CTS, but that is a large change.

Instead, an Annex B deframer is added to ESExtractor (trivial, but
tricky, code that should be in the parser anyway). Frame units are
provided, but the catch is that the OBU units within the frames are
also size-prefixed, as spec'd by Annex B, so bits of the annex B
parsing are kept, but names change to make clear their purpose.

Another issue with the Annex B parser in the NVIDIA SDK was that it
assumed OBUs would never have a size field, since they are redundant
with the sizing from the Annex B protocol. They are redundant, but the
spec allows them (providing they are consistent with Annex B). This
problem was discovered by some Argon test cases, which are very, very
strange.