THIS IS NOT THE REAL MAIN
Due to CTS requirements, external dependencies must exist on a branch
named main. For the samples, where CTS requires several changes that
aren't suitable for upstream, we are asked to fake it.

So this branch is very specific to the needs of the Vulkan CTS, and
has no relation to the main branch of the upstream project.